Build: Cover CSS-only ATTW exclusions
